test/mochitest/test_flash_b2g.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1235459
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="http://mochi.test:8888/chrome/browser/extensions/b2g-installer/test/mochitest/common.js"></script>
  <script type="application/javascript" src="http://mochi.test:8888/chrome/browser/extensions/b2g-installer/test/mochitest/flash_device.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<div id="container"></div>
<pre id="test">
  <script type="application/javascript" id="fake-adb"></script>
  <script type="application/javascript">

  "use strict";

  var deviceBlobfree;

  function continue_assert_finish() {
    test_unplugAdbDevice().then(() => {
      is(fakeFastbootDevice._flashed.length, 3, "Flashed three partitions");
      is(fakeFastbootDevice._flashed[0].name, "system", "Flashed system partition");
      is(fakeFastbootDevice._flashed[1].name, "boot", "Flashed boot partition");
      is(fakeFastbootDevice._flashed[2].name, "recovery", "Flashed recovery partition");

      let sizes = {
        zipFile:     1264564,
        bootImg:     497664,
        recoveryImg: 757760,
        systemImg:   16093632
      };

      return cleanupTmpDir(deviceBlobfree, false, sizes);
    }).then(() => SimpleTest.finish());
  }

  function runTest() {
    populateAboutB2GInstaller()
      .then(populateFakeDevices)
      .then(() => {
        fakeAdbDevice._files = { "/system/b2g/application.ini": "" };
        fakeAdbDevice._props["ro.build.type"] = "eng";
        fakeAdbDevice._props["ro.debuggable"] = "1";
        fakeAdbDevice._services["b2g"] = true;
      })
      .then(test_plugAdbDevice)
      .then(select_build.bind(this, 0))
      .then(zipFileName => deviceBlobfree = zipFileName)
      .then(assert_keepdata_visible)
      .then(click_flash);
  }

  setPrefsAndRunTests();

  </script>
</pre>
</body>
</html>
